Resolve cherry-picking conflict: move some conversion helper to `re_protos`
Note: doesn't need to be applied on `main
